refactor: move get daytona script to common
