+++ pkg/common/get_daytona_script.go
@@ -1,7 +1,7 @@
 // Copyright 2024 Daytona Platforms Inc.
 // SPDX-License-Identifier: Apache-2.0
 
-package constants
+package common
 
 import (
 	"strings"
