+++ pkg/common/get_daytona_script_test.go
@@ -3,7 +3,7 @@
 // Copyright 2024 Daytona Platforms Inc.
 // SPDX-License-Identifier: Apache-2.0
 
-package constants
+package common
 
 import (
 	"testing"
